[NTP Next] Refactor search box component
