+++ browser/resources/brave_new_tab_page_refresh/components/search/search_results.style.ts
@@ -3,18 +3,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import { color, effect, font, gradient } from '@brave/leo/tokens/css/variables'
+import { color, font, gradient } from '@brave/leo/tokens/css/variables'
 import { scoped } from '../../lib/scoped_css'
 
 export const style = scoped.css`
 
   & {
-    border-radius: 16px;
-    background: ${color.container.background};
     display: flex;
     flex-direction: column;
-    overflow: clip;
-    box-shadow: ${effect.elevation['01']};
   }
 
   .result-options {
